+++ module/item/item.js
@@ -4,3 +4,5 @@
  export class FSItem extends Item {
 
  }
+
+ //type: may not be undefined